@@ -29,10 +29,10 @@
     "tslib": "^1.9.0"
   },
   "peerDependencies": {
-    "@angular/common": ">= 7.2.16",
-    "@angular/compiler": ">= 7.2.16",
-    "@angular/core": ">= 7.2.16",
-    "@angular/forms": ">= 7.2.16",
+    "@angular/common": ">= 8.2.14",
+    "@angular/compiler": ">= 8.2.14",
+    "@angular/core": ">= 8.2.14",
+    "@angular/forms": ">= 8.2.14",
     "rxjs": "6.6.7"
   },
   "main": "bundles/ng2-semantic-ui.umd.js",
