@@ -52,7 +52,7 @@ export class SuiCheckbox implements ICustomValueAccessorHost<boolean> {
         return this.isDisabled ? "disabled" : undefined;
     }
 
-    @ViewChild("checkbox")
+    @ViewChild("checkbox", { static: true })
     private _checkboxElement:ElementRef;
 
     constructor() {
