@@ -53,7 +53,7 @@ export class SuiRadio<T> implements ICustomValueAccessorHost<T> {
     @Input()
     public isReadonly:boolean;
 
-    @ViewChild("radio")
+    @ViewChild("radio", { static: true })
     private _radioElement:ElementRef;
 
     public get checkedAttribute():string | undefined {
