@@ -56,7 +56,7 @@ export class SuiMultiSelectLabel<T> extends SuiTransition {
     }
 
     // Placeholder to draw template beside.
-    @ViewChild("templateSibling", { read: ViewContainerRef })
+    @ViewChild("templateSibling", { read: ViewContainerRef, static: true })
     public templateSibling:ViewContainerRef;
 
     constructor(renderer:Renderer2,
